Fix circular import after rebase
